pyproject.toml
[tool.poetry]
name = "dumpreporttoes"
version = "0.1.0"
description = ""
authors = ["sihyun <gthpgth@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
elasticsearch = "^7.17.0"
PyYAML = "^6.0"
beautifulsoup4 = "^4.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
